requirements.txt
pandas==2.1.1
scipy==1.6.2
streamlit==1.27.2
plotly_express==0.4.1